OneDrive/Desktop/railway-management-system/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
aiomysql==0.2.0
asyncpg==0.29.0  # For PostgreSQL support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Caching & Session
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Validation & Serialization
email-validator==2.1.0

# Development & Testing (move to requirements-dev.txt in production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing
factory-boy==3.3.0

# Production WSGI Server
gunicorn==21.2.0

# Environment
python-decouple==3.8